Create subdirectory view and move DialogBox and Ui files into it
