+++ src/main/java/duke/view/Ui.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.view;
+
+import duke.TaskList;
 
 public class Ui {
     private static final String GREET_MESSAGE = "Hello! I'm Duke \nWhat can I do for you?";
